Added file
This is a very broken version of the insulinReader. It is meant to be used as a pre-program for ELISA files coming off the plate reader. It is intended to make minimal assumptions, to allow future users to use this program as well. This program will ultimately not produce graphs, but produce tidy data files. These tidy data files can then be fed into another script that can combine multiple files based upon condition names.