Takes Condition Names
This is the 'tower of cards' release - as in, if you try to touch it or use it for anything useful, it will break. I hope to make it usable within the next few releases - there are still a few edge cases that haven't even been considered, much less stress tested.